requirements.txt
whisperx
torch
torchaudio
torchvision
ffmpeg-python
openai
moviepy
pyobjc